feat/#45: 리뷰 상세조회 API 연동

+++ app/src/main/java/com/konkuk/arabyte_aos/domain/model/ReviewDetail.kt
@@ -7,7 +7,7 @@ data class ReviewDetail(
     val reviewId: Int,
     val companyName: String,
     val isCertified: Boolean,
-    val star: Float,
+    val star: Int,
     val region: String,
     val category: ArabyteJobCategory,
     val reviewRating: ReviewRating,
